refactoring completed

+++ src/application/use-cases/note/delete-note-use-case.ts
@@ -1,14 +1,36 @@
-import NoteRepository from "@application/repositories/note-repository";
-import { NoteEntity } from "@domain/entities/note.entity";
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from '@nestjs/common';
+import NoteRepository from '@application/repositories/note-repository';
+import type NoteEntity from '@domain/entities/note.entity';
+import PasteRepository from '@application/repositories/paste-repository';
 
 @Injectable()
 export default class DeleteNoteUseCase {
-    constructor(
-        private readonly _noteRepository: NoteRepository
-    ) {}
+  private readonly _noteRepository: NoteRepository;
+  private readonly _pasteRepository: PasteRepository;
 
-    public async handle(pasteId: string, noteId: string): Promise<NoteEntity> {
-        return this._noteRepository.delete(noteId)
-    }
+  public constructor(noteRepository: NoteRepository, pasteRepository: PasteRepository) {
+    this._noteRepository = noteRepository;
+    this._pasteRepository = pasteRepository;
+  }
+
+  public async handle(pasteId: string, noteId: string): Promise<NoteEntity> {
+    if(await this.isValidNote(noteId) || await this.isValidPaste(pasteId)) throw new NotFoundException();
+
+    const response = await this._noteRepository.delete(noteId);
+
+    return response;
+  }
+
+  private async isValidPaste(id: string): Promise<boolean> {
+    const paste = await this._pasteRepository.getById(id);
+
+    return paste !== null;
+  }
+
+  private async isValidNote(id: string): Promise<boolean> {
+    const note = await this._noteRepository.getById(id);
+
+    return note !== null;
+  }
+
 }
